pyproject.toml
[project]
name = "paper-mind"
version = "0.1.0"
description = "Personal research mentor: ingest PDFs, build knowledge graphs, study with AI agents"
requires-python = ">=3.11"
dependencies = [
    "nicegui>=2.0",
    "litellm>=1.40",
    "python-dotenv>=1.0",
]

[project.scripts]
papermind = "main:main"
